refactor: to fix typo
